feat: userの設定変更を追加
