+++ src/pages/setting.tsx
@@ -3,7 +3,7 @@ import { User, Link, Notification } from "@prisma/client";
 import { GetServerSideProps } from "next";
 import { getServerSession } from "next-auth";
 import { authOptions } from "./api/auth/[...nextauth]";
-import { UserToLinkQuery } from "@/utils/query/Article.query";
+import { UserToLinkQuery } from "@/utils/query/User.query";
 import { HomeLayout } from "@/components/organisms/HomeLayout";
 import { Setting } from "@/components/template/Setting";
 import { NotificationQuery } from "@/utils/query/Notification.query";
